DRY up util of getting coordinate center
